Fix minor sunburstchart-bug
If the EventHandler is wrapped in the WeakEventHandler nothing prevents the EventHandler from being garbage collected. Therefore, the TreeNodeEvents  will no longer be fired at some random point in time.
However, removing the WeakEventHanldler should not result in memory-leaking, because if the chart is redrawn all segments are cleared, the path will get garbage collected and so the EventHanlder.