Added an experimental method to render to an image and save it
Hopefully all of the charts now have a method render the chart to an image. The method takes the image filename without ending and the width and height of the final image as parameter. The image will be rendered as a png file so that you will have a transparent background.